Remove runtime generation of server-specific openapi docs * Create separate resources for each server type to return their specific openapi doc * Reintroduce previous functionality in preparation for generation of openapi docs at build time
