+++ tessera-jaxrs/common-jaxrs/pom.xml
@@ -60,6 +60,7 @@
                 </exclusion>
             </exclusions>
         </dependency>
+
     </dependencies>
     
     <build>
