+++ tessera-jaxrs/common-jaxrs/src/main/java/com/quorum/tessera/app/TesseraRestApplication.java
@@ -1,6 +1,5 @@
 package com.quorum.tessera.app;
 
-import com.quorum.tessera.api.common.ApiResource;
 import com.quorum.tessera.api.common.BaseResource;
 import com.quorum.tessera.api.common.VersionResource;
 import com.quorum.tessera.api.exception.*;
@@ -30,9 +29,10 @@ public abstract class TesseraRestApplication extends Application implements Tess
                 WebApplicationExceptionMapper.class,
                 NodeOfflineExceptionMapper.class,
                 VersionResource.class,
-                ApiResource.class,
                 BaseResource.class);
     }
+
+    // TODO(cjh) return full swagger doc by default in case specific ApiResource method for that server hasn't been added
 
     @Override
     public CommunicationType getCommunicationType() {
