+++ tessera-jaxrs/transaction-jaxrs/src/main/java/com/quorum/tessera/q2t/Q2TRestApp.java
@@ -9,10 +9,12 @@ import com.quorum.tessera.service.locator.ServiceLocator;
 import com.quorum.tessera.transaction.EncodedPayloadManager;
 import com.quorum.tessera.transaction.TransactionManager;
 import com.quorum.tessera.transaction.TransactionManagerFactory;
-// import io.swagger.annotations.Api;
 
 import javax.ws.rs.ApplicationPath;
 import java.util.Set;
+import java.util.stream.Stream;
+
+import static java.util.stream.Collectors.toSet;
 
 /**
  * The main application that is submitted to the HTTP server Contains all the service classes created by the service
@@ -47,11 +49,17 @@ public class Q2TRestApp extends TesseraRestApplication {
 
         TransactionResource transactionResource = new TransactionResource(transactionManager);
         RawTransactionResource rawTransactionResource = new RawTransactionResource(transactionManager);
-        EncodedPayloadResource encodedPayloadResource
-            = new EncodedPayloadResource(encodedPayloadManager, transactionManager);
+        EncodedPayloadResource encodedPayloadResource =
+                new EncodedPayloadResource(encodedPayloadManager, transactionManager);
         final UpCheckResource upCheckResource = new UpCheckResource(transactionManager);
 
         return Set.of(transactionResource, rawTransactionResource, encodedPayloadResource, upCheckResource);
+    }
+
+    @Override
+    public Set<Class<?>> getClasses() {
+        return Stream.concat(super.getClasses().stream(), Stream.of(ApiResource.class))
+            .collect(toSet());
     }
 
     @Override
